WT-13389 Implement block_session extent unit tests (#10963)
This tickets aims to test the block manager session part of the code.
The idea behind the unit testing allows us to drill down into core
functions of the block session code and check it's behaviour. There are
a few technical changes I needed to make and would like to get reviews
on:

- I configured the s_functions.py file to allow __wti functions to be
tested outside of the module.
- I needed to move the WT_BLOCK_MGR_SESSION outside of the source file.
I used #ifdef definitions to do this if HAVE_UNIT is enabled.
- I created unit test functions version of all the internal functions
used inside block_session.c
- I have added a block manager session creation inside the
mock_session.c, that is used for almost all test cases.
- I have added a new unit test file that aims to test the core functions
inside block_session.c

**Update**: This PR will only test block session ext functions.