WT-13451 Update free block functions to pass in double pointers and make it consistent with wiredtigers free pattern (#10988)
The ticket changes the free functions to pass in double pointers
instead. The reason behind this change is for testing purposes. When we
pass in a single pointer reference that is going to be freed, C does a
copy-by-value of the pointer and does not clear the original pointer
reference. Therefore if we pass in the pointer to the pointer, we can
properly clear the original pointer to NULL.