WT-13403 Implement block manager file open unit tests (#10989)
This tickets aims to test the block manager flile open. There are a few
technical changes I needed to make and would like to get reviews on:

- I have created a config parser that deals with configuring wiredtiger
configurations. It would construct the config string and pass it into
the first entry in the char ** array.
- I have tried to test all the possible scenarios in the file open.
- There are a few scenarios where it is not possible to test. For
example, the function creates a local flag that is not persisted
anywhere inside the block struct. I will create a new ticket to address
the testability of this. Another case requires the file size to branch
into the if statement. This was proven hard to test.
- I currently use `__wti_bm_close_block` to clear the block references.
- I created a setup in the mock connection and destruct it in the mock
session and mock connection. I added a destruct in the mock session
because it requires a reference to the session.