WT-13512 Fix WT_SESSION::reconfigure when logging is enabled (#11015)
This fixes an issue in `__config_merge_scan`, which did not recognize
key/value pairs such as `log=(enabled)` to be recognized as structs.
This had the unfortunate consequence of merging `log=(enabled=false)`
and `log=(enabled)` to still result in `log=(enabled=false)`.

This could, for example, cause `WT_CONNECTION::reconfigure` to not work
if the connection was opened with just `log=(enabled)`. In that case,
the connection's own config string would still contain `enabled=false`
but the connection flags would have `WT_CONN_LOG_ENABLED` set. This
mismatch would then cause `__wti_logmgr_config` to fail.