Remove use of free_space_target parameter and outdated FIXME
free_space_target is not available on 7.0 and earlier. We can't use it to guarantee compaction takes place, intead use a larger number of keys to make sure we hit the free_space threshold for compaction to trigger.

Also remove the WT-11399 FIXME as work on tiered storage has stopped.