WT-13546 Use internal session to create the history store table (#11028)
The connection's default session is used to save/restore the connection
state whenever a connection API is called. The same default session is
also used for creating the history store table. After allowing the
statistics cursor to be opened during recovery, it is possible that the
stats collector thread can interact with the default session while it is
creating the history store table.

To avoid these unnecessary interactions with the default session, use
the internal history store session while creating the history store
table.