WT-13426 Create block manager file read/write API tests (#11021)
This tickets aims to test the block manager API operations
read/write/write_size. There are a few technical changes I needed to
make and would like to get reviews on:

- The test cases that are written, if I am missing anything
- Creation of the handle because WT_ASSERT requires it within
wti_block_alloc.
- How the test cases are structured, can it be improved?

**Update**
- In this pull request I have converted to using from an inmemory file
system to utilise the local filesystem. This required me to change files
outside of the added file. The reason behind this is because there are
specific functions that are not being tested because inmemory filesystem
was used.

- There were memory bugs with regards to the block manager session
cleanup function, since implementing write test cases. I have properly
fixed it now.

---------

Co-authored-by: Luke Pearson <luke.pearson@mongodb.com>