LABS-1004 Page and Log Interface and PALM prototype implementation. (#11045)
This branch implements a mockup of a page and log interface (PALI) as described in LABS-849.
The mockup is called PALM, and uses lmdb (a small kv store that is part of open ldap).
There is a single test that drives it, showing that it can store full pages and deltas and then retrieve
them. PALM is just a simple implementation of PALI, useful for testing and development.