WT-13561 Refactor Eviction Public/Private API (#11050)
Now that all the eviction-related files are present in the evict/ folder
and there is folder-module mapping, it becomes easier to use the current
prefixes as they were originally intended. In this PR, I have done the
following:

1. Removed functions (mainly public APIs) that were not part of eviction
-
Functions that were irrelevant to eviction have been removed to clean up
the API - `__wt_page_release_evict` and
    `__wt_page_dirty_and_evict_soon`
3. `__wti_btree_dominating_cache` function is only used in evict_lru.c,
so it was relocated to improve modularity.
4. Renamed functions with prefixes starting with __wt_evict in the evict
folder (except verbose functions) - After discussions with Etienne, we
decided not to rename verbose-related functions yet, as a separate
module will be created for verbose. Renaming them now would create
unnecessary inconsistencies, so I have left them as-is.