WT-13521 POC-weaken-session-log-dependency-2 (#11035)
This is an option to reduce the dependency between the session module
and the log module.

session invokes __wt_log_op before logging the transaction operation.
The usage is as :

```
   if (__wt_log_op(session))
      WT_ERR(__wt_txn_log_op(session, NULL)); 
```

All the other callers of __wt_log_op have the same above usage pattern.

__wt_log_op indicates whether a transaction operation needs to be logged
or not.
__wt_log_op is an inline function.

For performance reasons, it is better to have the check (__wt_log_op) as
inline, and based on the check result, a relatively heavy function
__wt_txn_log_op will be invoked.

The thought behind this change suggestion is to, the conditional check
to determine whether an operation is required and whether the operation
itself, should belong to the same module.
In this case, the transaction module should be responsible for
abstracting the functionality to determine whether transaction operation
logging is needed or not.
log module is the transaction log module, my view is that if the
transaction module doesn't exist, then the log module also doesn't
exist. i.e. the transaction module is composed of the log module.
In this case, I prefer the dependency of other modules like the session
module can be on the transaction module, rather than directly on the log
module which is part of the transaction module.

This change, I am not anticipating any performance regressions, it is
just about a module dependency on two different modules 1) transaction
and 2) log, the dependency is reduced to only one module 1) transaction.

Before this PR, the dependency is :

```
Session --> Transaction
Session --> Log
Transaction --> Log
```

With this PR, the dependency becomes :

```
Session --> Transaction
Transaction --> Log
```