LABS-970 Add page IDs and page metadata relevant to the block storage (#11053)
This PR adds a metadata struct to `WT_PAGE`, called
`WT_PAGE_BLOCK_META`, which stores information that is relevant to how
the page is stored. It currently stores only the page ID, but it will be
expanded in the future to track other relevant storage metadata, such as
the base checkpoint ID, number of deltas since the last checkpoint, etc.

How it works:
* `WT_PAGE_BLOCK_META` is stored on the page struct itself. It is
initialized together with the page, and it it is filled out by the Block
Manager during page read. We expect the page read interface will provide
info such as base checkpoint ID and the number of deltas in the future.
* The block metadata struct `WT_PAGE_BLOCK_META` gets passed on the path
from page to the Block Manager, which then uses it to get the pantry ID
during write, and which fills it out when reading a page.
* Reconciliation uses the `WT_PAGE_BLOCK_META` struct from the page
itself when writing a replacement page, which results in writing a page
with the same ID. In the future, we would need to change this to keep an
updated copy of this struct with metadata from the last reconciliation.
This should be in another ticket; possibly with the checkpoint ID
ticket, as we will incorporate new checkpoint IDs during a write.
* If splitting a page, the first page currently uses
`WT_PAGE_BLOCK_META` from the original page, while the subsequent pages
use a new struct with a new page IDs. If we'd like we can change this to
always using new page IDs for splits in the future. We would also need
to keep along these structs that we used in the previous splits; this
should come later as we implement page deltas for splits.
* The block cache keeps a copy of `WT_PAGE_BLOCK_META` alongside the
cached data chunk.
* The page ID allocation is contained within the b-tree itself, and the
next available page ID is persisted in the b-tree's checkpoint metadata.