WT-13519 Add __wt_lsn_file and __wt_lsn_string to the logs public API (#11056)
Add a `__wt_lsn_file` function to access the `WT_LSN::file` field. This
pairs with the existing `__wt_lsn_offset` function to access
WT_LSN::offset.

Also add a `__wt_lsn_string` function converts and LSN to a raw `%d,%d`
string format. A side effect of this change is the meta/ module no
longer depends on log/

There's a large number of direct accesses to lsn.l.file inside the log
module. I've decided to leave those accesses alone for now and only use
the new function outside of the log module. It seems reasonable for a
module-private field to be accessed directly inside the module.