WT-13442 Do not allow out of order truncate commit timestamp (#11069)
WiredTiger doesn't allow out-of-order commit timestamps and it
return an error whenever an operation is trying to commit with
a timestamp less than the existing timestamp on an update.
However, it failed to verify the truncate commit timestamp.

The WiredTiger timestamp usage check is extended for the truncate
operations also doesn't allow an out of order commit timestamp
to be performed.