WT-13613 Add modularity_check tool (#11091)
Add a tool for developers to determine how modularized a component is in
WiredTiger. It can report who uses a module, who a module uses, report
dependency cycles in the code, and provide a full summary of what
structs and functions are private to a module.

Further details are provided in the `README.md`