LABS-1053 Make the ingest table an in-memory btree (#11070)
Rename the cache_resident configuration to use another flag
WT_BTREE_NO_EVICT. We cannot use the cache_resident configuration for
the ingest table as it prohibits running eviction on the tree.

Add another option to create an in-memory btree for non-in-memory
database. We will create a file for this btree in this case but we never
write anything to it. For the in-memory btree, we reconcile its pages
the same way as the btrees in in-memory database during eviction. While
checkpoint always skips this type of btrees.

Use the in-memory btree for the ingest table. The ingest table cannot
write data to the disk as it shouldn't insert updates to the history
store.

The reconciliation of the in-memory btree can be used as the garbage
collection mechanism for the ingest table.