WT-13628 POC: Small improvements to make using Control points easier, renaming for consistency, and fix error handling (#11111)
The main reason for this change is to add the macros and needed for
using predicate "Parm 64 Match".
Without the macros usage of "Parm 64 Match" requires lots of ifs to
handle control_points NULL and cp_data NULL.

This change also fixes error handling. Returning WT_ERROR causes an
abort. Therefore return the correct error, not that.

This change has the most important renaming. Other renaming is not
included.
* Rename predicate "Param match" to "Param 64 match".
* Rename action "Wait for trigger" to "Trigger".
* Rename the control point define and control point call macros.