WT-13478 Improve Block manager test consistency (#11099)
This ticket and PR improve the consistency of the Block manager
Catch2-based tests. In particular, tests have been refactored and helper
code moved to be clearer to understand and make better use of Catch2.

Changes make by this PR:
- Moved shared block manager test utility functions into
`block_util.cpp`
- Separated TEST_CASE for __wti_block_ext_prealloc into two TEST_CASES
to simplify test setup/cleanup
- Refactored extent list tests to use common setup/cleanup where
appropriate
- Changed `operator()(const off_size &left, const off_size &right)` to
be `const` as it can be.
- Tidied up constness for function parameters.