WT-13502 Add 'backup_checkpoint' to 'query_timestamp'. (#11011)
Add a configuration setting to be able to query the timestamp that is
the timestamp of the checkpoint pinned by an open backup cursor.

@luke-pearson and @keitharnoldsmith here is an implementation of what we
discussed. It basically works but I didn't go into situations where
timestamps aren't in use or checkpoints sometimes taken without the
stable or weird cases like that. This is the basic functionality and
plumbing. But I did make a test!