LABS-1117 Merge the code to write leaf page delta (#11108)
Here is how the leaf page delta works currently:

* Init a tombstone on the update chain if there is a stop timestamp on
the on-disk key when reading the page into memory. This is used to track
that whether we have deleted the key in the delta. If we don't do that,
we may never delete that key in the delta.
* Mark the update that has been reconciled using a flag
WT_UPDATE_DURABLE. We only consider the update that is not marked with
this flag in the delta.
* We build the delta if we don't split the page and there is an existing
in-memory disk image (This is too strict but it is here for simplicity
now).
* The way we build the delta now is to save all the update chains that
we may need and iterate them later after we have built the disk image (I
do this after we have built the disk-image because I want to save the
work if we find out that we cannot build or should not build the delta
in the end).

---------

Co-authored-by: Etienne Petrel <77254506+etienneptl@users.noreply.github.com>