WT-13646 Fix link issue with catch2-unittest-assertions (#11122)
Moved test_and_validate_write_size() to util_block.cpp to fix link
issues that occur when HAVE_UNITTEST_ASSERTS is set.

Note that it was necessary to add a parameter to
test_and_validate_write_size() to pass in the allocation size rather
than rely on a file-scope value.