WT-12644 Use normalized soft position to remember positions in Eviction Server (#10924)
This PR uses the "soft position" when the pointer to a page needs to be
stored for long time. Each time when there is a chance that Eviction
Server will sleep, all hazard pointers held by eviction are translated
into "soft positions" and released. Next time, when Eviction Server
wants to continue its walk, it restores the position at approximately
where it left off.

Because pinned pages are no longer an issue, this PR removes the
restriction on the number of hazard pointers used by Eviction Server in
one pass (`WT_EVICT_MAX_TREES`) and removes the corresponding metric
`cache_eviction_server_skip_trees_too_many_active_walks`.

This PR adds logging of saved and restored positions at the level
debug_1.

A counter is added for each time when restored position doesn't match
the page for which it was saved.

Unit test makes sure that the page restored from the saved position is
the same one (as long as the tree is not modified).

This PR tries to **keep** everything about how Eviction Server walks the
tree.

Performance tests:
*
https://performance-analyzer.server-tig.prod.corp.mongodb.com/perf-analyzer-viz/?comparison_id=5cd3d9cd-c53a-456f-8677-407a5b0ce845
* https://spruce.mongodb.com/version/66df742c354f28000731dffb

Doc:
https://docs.google.com/document/d/17X0MkeVoRELaQuEfSvIbVBgCZqKcWg-GSf_Enth6Ovk