WT-13642 Add separate statistics for dirty leaf and internal page bytes (#11141)
We currently expose the tracked number of dirty bytes in the cache as a
statistic: `cache_bytes_dirty`
However, this does not differentiate between leaf or internal pages.
This metric is currently incorrectly interpreted in t2 as the reason for
the cache being in the `WT_EVICT_CACHE_DIRTY_HARD` state and causing
application threads to evict. This is misleading because the dirty hard
state is set based on the dirty leaf page bytes not the total dirty
bytes.

The dirty leaf and internal bytes are already tracked in the cache
structure as:
`bytes_dirty_intl`
`bytes_dirty_leaf`

This PR exposes these as new statistics.