WT-13555 Initalize the statistic framework into mock connection for better unit testability (#11167)
In unit testing whenever a WT_CONN_STAT_INCR is invoked, we ignore it as
in unit testing statistics are not enabled as part of wiredtiger_open.
In WiredTiger statistics are incremented when certain code paths are
hit. Unit tests can assert that certain statistics are incremented when
expected code paths have been executed.

This ticket involves improving the mockConnection class by initializing
the statistic framework.