WT-13710 Fix "s_docs -l" run failure (#11184)
The "s_docs -l" command run failed because the `doxyfile` variable is
not set. The setting is located in the `build` function, which is not
called in the command invocation path.

The solution is to abstract the doxyfile setting and doxygen command
checking logic into a new function, and call the function in all paths
that doxyfile is needed. Also added 2 symlink files to avoid the
doxyfile not found error when running the command with the specific
versions of doxygen (9 and 11).