WT-13525 Migrating to AWS-based MacOS hosts (#11174)
DevProd are migrating our MacOS hosts from MscStadium to AWS so we are
making changes in WT evergreen configuration to switch from
macos-13-arm64 to macos-14-arm64. Our compilation requires the swig
package, which is no longer installed in macos by default. Therefore I
have modified the script for configure_wiredtiger and make_wiredtiger to
install swig as a temporary workaround. A ticket has been filed to
install siwg on our macos-14-arm64 distro and the temporary workaround
will be removed after that ticket is completed.