WT-13612 Fix split scenario that would lead to a time aggregate having incorrect values (#11158)
When splitting a page, sometimes WiredTiger will merge part of the
`prev_ptr` chunk into the `cur_ptr` chunk. The accounting for time
aggregates was wrong in this scenario and previously `cur_ptr` would be
left with a time aggregate that is incorrect. Fortunately, due to how
time aggregate merging works this is safe as the values chose when
merging the `prev_ptr->ta` into the `cur_ptr->ta` would always be the
more conservative of the two.