SLS-508 Fix cache statistics (#11191)
SLS-8 turned these off since it was mildly more tricky with deltas. But
that breaks MongoDB's testing, which looks at a couple of cache stats
and expects them to move in a certain way. This caused a test failure
which prevented the WT drop. Reinstate them, and fix them for deltas.