WT-13716 Prevent retention of fast truncated history store pages (#11190)
The problem described in the ticket is that, since the HS is
non-transactional, page_del->committed is not set. This prevents the
deletion of the page, causing reconciliation to retain the original
version. By treating HS pages that are fast-truncated during selective
backup as globally visible (even though they are non-globally visible)
and setting page_del to null, we eliminate this issue. This allows us to
remove the pages without checking visibility information, which is
acceptable because the HS is non-transactional.