SLS-51 Write disk image for update restore eviction (#11215)
We currently skip write the disk image to disk if it's an update restore
eviction because it will be written again by the checkpoint later
anyway. However, it is not correct if we have delta. We need to ensure
the delta is built on a disk image that has been written to disk. Change
to always write the disk image even for update restore eviction.