WT-13717 Allow variables to be declared other than at the start of a C function (#11195)
This change updates the WiredTiger C coding standard to allow and
recommend, but not require, that local variables are declared at the
smallest scope and are initialized when declared. This brings it in line
with the WiredTiger C++ guidelines.

The change also updates the compiler flags to match the updated coding
standard, including ensuring there are checks for `goto`s that jump over
variable initialisation, and updates/removes checks in `s_all` to
correspond to the updated coding standard.

Also includes are a few simple examples of applying the updated coding
standard to existing code.