Always set the WT_ITEM_ALIGNED flag for scratch buffers.
It's possible for code to do a WT_CLEAR on the item, so when it gets freed
and returned to the scratch buffer pool, it will still have that flag unset.
Unfortunately, we previously only set the flag when we allocated a WT_ITEM,
so cached items would remain unaligned.

Only one of the two affected tests is properly fixed, as test_oligarch07
hits another bug that's being investigated in a different branch.

This change also removes a leak in `__wt_disagg_get_meta` where we replaced
the contents of an allocated item.