SLS-494 Store the metadata for shared tables in a shared metadata table (#11211)
Store metadata for the stable oligarch tables in a shared metadata
table.

How it works:
* Store the metadata of stable oligarch tables in a new shared table
called `WiredTigerShared.wt_stable`.
* Store the metadata of `WiredTigerShared.wt_stable` itself in a special
"table" in PALI, akin to the turtle file.
* When picking up a checkpoint, get the latest version of
`WiredTigerShared.wt_stable`, and then apply the metadata from it to
`WiredTiger.wt`.

This wouldn't scale for a large number of tables, but it should be good
enough for the demo. Some ideas about how to scale this are in the
design doc; we'll do that later.

Additionally:
* Add printing table IDs to PALM's verbose logging.
* Add extra checkpoints and delays to `test_oligarch09` to ensure that
(1) all data are processed by the watcher before the corresponding
checkpoints, and (2) we know the correct checkpoint ID from which to
reopen the connection.
* Add a new test `test_oligarch13` that uses multiple tables, both
oligarch and shared non-oligarch.