SLS-518 Fix largest key implementation and implement a simple random cursor for oligarch table (#11239)
Fix the implementation of the largest key implementation. The previous
implementation only implements the largest visible keys. However,
largest key should return the largest key even it's not visibility.

Implement the random cursor function for oligarch table. I borrowed the
main implementation from the lsm code. It is not a uniform random
solution yet as it doesn't consider the key size of the ingest table and
the stable table but it should work for now.

Fix a bug in oligarch search near.