WT-9102: update to macos 11
This commit bumps our builds from MacOS 10.14 to MacOS 11. 10.14 has been
unsupported for several years and will be removed from Evergreen soon.
MacOS 11 is also EOL, but I didn't want to move all the way to MacOS 14
on an older branch until 11 is also removed from Evergreen.

The only significant changes I had to make here were altering the
python_binary and adding -Wno-poison-system-directories to the non-cmake
build. This flag seems to already be used for MacOS cmake builds, so
I figured it was alright to add for the non-cmake build as well.