WT-13840 Remove and rename unused symbols (#11293)
Now LSM is removed, there are a few functions in the cursor and bloom
modules that are no longer used outside of their file or module. These
are renamed to remove the `__wt` prefix or to use the `__wti` prefix.

Some macro usage also changed:
`WT_SESSION_LOCKED_HANDLE_LIST_READ` is now no longer used explicitly
but is used as part of another macro `WT_SESSION_LOCKED_HANDLE_LIST`. It
is therefore excluded from `s_define`.
`WT_TIMECMP` is no longer used, so it is removed altogether.