WT-13835 Updating the live restore API configuration (#11295)
The PoC live restore configuration relied on an `aux_path` parameter to
specify the live restore path, additionally the existence of a path
indicated that a live restore was in progress. A better defined API is
needed to distinguish live restore from other usages of `aux_path`. To
that end the `aux_path` config is removed and a `live_restore`
configuration takes its place.

It accepts the following options:
- `enabled`
- `path`
- `threads_max`