WT-11946 Bug Fix for the WT command line tool JSON dump option (#11307)
This error occurs when the WT command line tool is used to dump a table
in json format (with the `-j` param).

There is a buffer (`buf`) in `json.c` being reused for loading lines of
text to be dumped, and in some cases (particularly if subsequent lines
are shorter than previous ones) the buffer is underfilled without adding
a corresponding null terminator.
This leads to junk values being included at the end of each line in the
dump output.

I've added in that null terminator where it's needed, as well as a check
that the `bufsz` counter was calculated correctly, such that when it's
time to add the closing quotes and null terminator, `bufsz = 2`.

I've also written a test case that creates a table with key/value pairs
of varying lengths (this would guarantee the inclusion of junk values in
the dump, pre-bugfix) and validates the dump output.

---------

Co-authored-by: Zunyi Liu <43444919+zunyiliu@users.noreply.github.com>