WT-13549 Investigate block checksum getting cleared when performing read() (#11304)
After investigating the code logic, it is clear that blk->checksum needs
to be cleared for the calculation of the checksum and will remain
cleared since it will not be revisited during a B-tree traversal.
Therefore, there is no need to reset it to its original value, and no
code changes are required.

However, during the investigation, a lot of time was spent understanding
the logic of the code and why blk->checksum is cleared and not reset. To
avoid similar confusion in the future, I am adding a comment here for
reference.