WT-13834 Move __wt_ckpt and __wt_ckpt_snapshot to checkpoint.h (#11290)
Move `__wt_ckpt` and `__wt_ckpt_snapshot` to `checkpoint.h` and change
prefixes for everything that's moved.

`__wt_block_mods` is moved too as it's a part of `__wt_ckpt`.

`WT_CKPT_BLOCK_MODS` is renamed to `WT_CKPT_BLOCK_MODS_LIST` because of
the name clash with `__wt_block_mods` which is renamed to
`__wt_ckpt_block_mods` to comply with the naming convention.