WT-11943 Bug Fix for WT command line tool dump command with -j -k options (#11306)
An error occurs when the WiredTiger command line tool dump command is
used with the -j -k options.

The error occurs because the dump cursor is in json format mode, which
requires keys to be formatted in json to be set. For example, if the key
is ‘name’, to set the key it must be formatted to be ‘“key0” : “name”’.

I've added a json wrapper to the key before setting the key on the dump
cursor. It will only format the key when when both -j and -k are used.
I've also added an additional test file which tests the use of -j and -k
together.

(cherry picked from commit 3c249eaf0c3fdf64b387f4a39a1f21d98b66918d)