WT-13860 Fix timestamp check on wrong prepared update (#11330)
The assumption in the code flow that is wrong is during the prepare
update resolution by a transaction, the update chain will either have
the aborted updates or only the prepared updates by that transaction.

All the prepared updates from a transaction are resolved as part of the
resolution of the first operation on that key and the subsequent
operation resolution on the same key will be no-op. Hence even when the
commit of a transaction with multiple operations on the same key is in
progress the prepared updates of that transaction in the update chain of
a particular key could be resolved and hence new prepared updates by
other transactions could be added to the update chain.