WT-13837 Set up code owners to guard backport PRs (#11340)
The `.github/CODEOWNERS` file introduced in this PR coupled with a newly
created
[backport-approvers](https://github.com/orgs/wiredtiger/teams/backport-approvers/members)
Github team will enable one member of the team to be auto-assigned as a
reviewer for backport PRs in a round-robin fashion to guard the PR
merge, i.e. backport PRs without approval from one of backport-approvers
won't be able to merge. This is in line with the [server backport review
& approval
policy](https://wiki.corp.mongodb.com/display/KERNEL/Database+SERVER+Backports+Policy+and+Workflow#DatabaseSERVERBackportsPolicyandWorkflow-Approving/DecliningaBackport).