WT-13861 Create placeholder function API and error struct (#11327)
This ticket is part of the project to improve error reporting and
handling between WiredTiger and MongoDB
([SPM-3810](https://jira.mongodb.org/browse/SPM-3810)).

I have created a placeholder API at the session level,
`__session_get_last_error`, that simply returns error and sub-level
error codes of `0`, along with a placeholder error string. These will be
replaced with actual default values in
[WT-13863](https://jira.mongodb.org/browse/WT-13863).

I also created a new struct, `WT_HELPER_ERROR`, to be used by the
session to store information about the most recently occured error.

This change required modification of the SWIG interface (`wiredtiger.i`)
for the Python API to work with the new function (using char * as an out
parameter).

_Note: The originally proposed name for the struct, `WT_VERBOSE_ERROR`,
was found to be already used in the code as an enum value. A new name is
yet to be confirmed, once it has been all documentation should be
updated accordingly._