WT-13642 Add separate statistics for dirty leaf and internal page bytes (#11141) (v8.0 backport) (#11354)
We currently expose the tracked number of dirty bytes in the cache as a
statistic: `cache_bytes_dirty`
However, this does not differentiate between leaf or internal pages.
This metric is currently incorrectly interpreted in t2 as the reason for
the cache being in the `WT_EVICT_CACHE_DIRTY_HARD` state and causing
application threads to evict. This is misleading because the dirty hard
state is set based on the dirty leaf page bytes not the total dirty
bytes.

The dirty leaf and internal bytes are already tracked in the cache
structure as:
`bytes_dirty_intl`
`bytes_dirty_leaf`

This PR exposes these as new statistics.

(cherry picked from commit ca9ffe8305d5815fd44b2cd1ff3c3fc6380c6984)

Summarize the reason behind this change (this might be the problem
you're solving, or the context around the request) and the solution you
have chosen.

Co-authored-by: Sean <41533874+Sean04@users.noreply.github.com>
(cherry picked from commit 1416a711395d191efac39820ab59697b4f51942c)