WT-13911 Fix configuration related issue in many-collection-test (#11351)
Recently the many-collection-test Evergreen task has had a number of
failures, where the test has failed to find the mongod binary. It seems
that (sometimes at least) a symlink is used to point to the actual
location of the mongodb binary, and the code in evergreen.yml was not
finding the binary as a result.

The solution is to add the '-L' command line option to `find` to follow
symbolic links. This change also prints out the mongod path to assist
with future debugging.