WT-12846 Fix how compact walk handles EBUSY from checkpoint flush_lock (v8.0) (#10929) (#11359)
The compact tree walk requires the checkpoint `flush_lock` while it
walks internal pages deciding what pages to rewrite. We currently use
trylock to retrieve the lock and give up the walk if we can't get it.
This prevents us from walking the whole tree when traversing the tree
takes longer than the period between checkpoints.

This moves the handling of EBUSY returned from retrieving the flush_lock
from the session compact layer where we count the passes over a file to
the btree layer where we traverse internal pages. This has the benefit
of not abandoning the tree walk as well as reusing the interrupt check
and cache stuck check that we perform when moving between internal
pages.

(cherry picked from commit
https://github.com/wiredtiger/wiredtiger/commit/430071ff616b27374d88fa2958d955bbeaee6edb)