WT-13935 Move reconciliation flags from btmem.h to src/reconcile (#11362)
As we progress with reconciliation modularity, we wish to ensure that
symbols from elsewhere are moved to src/reconcile.

We have identified a set of flags that belong to src/reconcile but are
currently a part of btree module.

As part of this ticket we would aim to move these symbols to
src/reconcile.