SLS-818 Add new timing stats to both block managers (#11374)
This adds latency histograms to both block managers. Both block managers
use the same stats, since (1) that's consistent with the existing stats
and (2) I confirmed that's going to work for the analysis people want to
do.

The latency buckets are the same as for the filesystem, it remains to be
seen whether these are appropriate.