WT-13808 Add directory_open support to live_restore (#11379)
Add live restore support for opening file handles on directories. 

Directories behave differently to files. When opening a directory it is
immediately created in the destination so there's no concept of
background migrating a folder (files inside the folder are migrated
independently), and so no need for a hole list to be tracked.

This PR also adds some simple testing for `fs_file_open` of files and
directories, and fixes behavior when trying to open a file or folder
that:
- Does not exist in the source and destination, or
- Does not exist in the destination and has a tombstone
In both cases ENOENT will now be returned if WT_FS_OPEN_CREATE is not
set.

---------

Co-authored-by: Luke Pearson <luke.pearson@mongodb.com>