WT-13864 Capture message and error from WT_RET_MSG and WT_ERR_MSG (#11361)
An API has been created to get information about the last error that
occurred in a session, which should have been captured by certain RET
and ERR macros.
This change involves updating the `WT_RET_MSG` and `WT_ERR_MSG` macros
to track the last error and error message that has occurred in the
session API call, storing it in the session's `error_info` member.

---------

Co-authored-by: Dylan Liang <dylanthekid350@gmail.com>