WT-11253 Add basic ignore_prepare to test format. (#11358)
To add the basic capability to check ignore_prepare to test format.
This change sets the ignore_prepare setting to transactions randomly and
if ignore_prepare is set will ensure only reads are performed as per the
transaction semantics.

Changes:
1. function name change to reflect the function logic.
2. Occasionally set ignore_prepare and ensure the transaction will
perform only read operations as write operations are forbidden when
ignore_prepare is set.