WT-14008 Force reinstall layercparse when using modstat (#11439)
For some reason, pip install for layercparse sticks to the current
version, so new commits don’t trigger a fresh installation of
layercparse. To address this, we need to forcefully reinstall
layercparse whenever newer commits are made to the repo. This can be
achieved by appending --force-reinstall to the pip install command.