SLS-989 Remove layered log draining from layered datastores (#11440)
The initial implementation of layered datastores used the WiredTiger WAL
to move content from the ingest to stable table.

The current plan is to write the content directly into the stable table
on the primary, and to drain the ingest table on step-up via explicit
cursor traversal. Remove the code to tail the WAL and apply changes to
the shared table in favor of that solution.