WT-13946 Set sub-level error code and message with python tests for drop EBUSY workflows (#11475)
This ticket involves reproducing two of the EBUSY drop workflows and
making sure that the get_last_error API works as appropriate. The idea
is that the API should be able to provide more information on why an
EBUSY has been returned from the drop() call.
The following scenarios have been reproduced, handled using
`WT_RET_SUB`/`WT_ERR_SUB` macros, and tested:

- `EBUSY`, `WT_CONFLICT_BACKUP`, the table is currently performing
backup and cannot be dropped
- `EBUSY`, `WT_CONFLICT_DHANDLE`, another thread is currently holding
the data handle of the table