WT-14046 Fix python test for WT_CACHE_OVERFLOW sub-level error code (#11520)
Fixed a bug in the python test for the `WT_CACHE_OVERFLOW` sub-level
error code which caused it to sometimes fail on when ran on a windows
machine. This was due to a race condition between the application worker
thread and the eviction server, with the eviction server evicting data
from the cache before the application worker thread is able to check it.

The change consists of using a for loop and essentially rerunning the
original test multiple times to ensure that there are too many pages for
the server to evict all of them before the application worker thread is
able to check the cache.