WT-14005 Add live restore backup verification to format. (#11515)
Adding a live restore version of backup verification to test/format.
Currently have it configured 50% of the time.
With a stressful basic CONFIG and live restore turned on, it fails < 50%
of random configs. So I think it is finding bugs. I don't think it is a
bug in the format change.