SLS-1068 Write WiredTiger history file to SLS (#11534)
Write the history for shared tables into the shared storage.

The solution includes:
* Split the history file into two files, one for the history of local
tables and the other for shared tables. (Note, though, that the rest of
the code assumes that the history could be later split into a different
number of tables.)
* Annotate each instance of getting a HS cursor by the B-Tree ID, so
that the cursor open function selects the appropriate HS.
* Keep the B-Tree ID across HS cursor resets, because a different B-Tree
ID may necessitate a different HS cursor. Remove corresponding usages of
cached HS cursors that depend on clearing the B-Tree ID.
* For the few places that require iterating over the entire HS, provide
a lightweight mechanism to iterate over all available HS tables.
* A new test for timestamped reads after restarting without local files.